requirements.txt
bs4
requests
dotenv
yfinance
sentence-transformers
faiss-cpu
